+++ app/src/test/java/org/simple/clinic/drugs/EditMedicineEffectHandlerTest.kt
@@ -1,6 +1,8 @@
 package org.simple.clinic.drugs
 
+import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doReturn
+import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
@@ -9,7 +11,6 @@ import com.nhaarman.mockitokotlin2.whenever
 import dagger.Lazy
 import org.junit.After
 import org.junit.Test
-import org.simple.sharedTestCode.TestData
 import org.simple.clinic.drugs.search.DrugFrequency.BD
 import org.simple.clinic.drugs.search.DrugFrequency.OD
 import org.simple.clinic.drugs.search.DrugFrequency.QDS
@@ -20,13 +21,10 @@ import org.simple.clinic.mobius.EffectHandlerTestCase
 import org.simple.clinic.overdue.AppointmentRepository
 import org.simple.clinic.patient.SyncStatus
 import org.simple.clinic.protocol.ProtocolRepository
-import org.simple.clinic.storage.Timestamps
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequency
-import org.simple.sharedTestCode.util.TestUtcClock
 import org.simple.clinic.util.scheduler.TestSchedulersProvider
 import org.simple.clinic.uuid.UuidGenerator
-import java.time.Duration
-import java.time.Instant
+import org.simple.sharedTestCode.TestData
 import java.util.UUID
 
 class EditMedicineEffectHandlerTest {
@@ -35,7 +33,6 @@ class EditMedicineEffectHandlerTest {
   private val protocolRepository = mock<ProtocolRepository>()
   private val prescriptionRepository = mock<PrescriptionRepository>()
   private val facility = TestData.facility(uuid = UUID.fromString("94db5d90-d483-4755-892a-97fde5a870fe"))
-  private val utcClock = TestUtcClock(instant = Instant.parse("2018-01-01T00:00:00Z"))
   private val uuidGenerator = mock<UuidGenerator>()
   private val appointmentRepository = mock<AppointmentRepository>()
 
@@ -54,7 +51,6 @@ class EditMedicineEffectHandlerTest {
       protocolRepository = protocolRepository,
       prescriptionRepository = prescriptionRepository,
       facility = Lazy { facility },
-      utcClock = utcClock,
       uuidGenerator = uuidGenerator,
       appointmentsRepository = appointmentRepository,
       drugFrequencyToLabelMap = drugFrequencyToLabelMap,
@@ -71,7 +67,6 @@ class EditMedicineEffectHandlerTest {
   @Test
   fun `when refill medicine effect is received, then clone the prescription and save it in the repository`() {
     // given
-    val durationToAdvanceBy = Duration.ofMinutes(20)
     val patientUuid = UUID.fromString("7bb4616b-1542-46ad-93d5-9d03b619aa99")
     val facilityUuid = UUID.fromString("32dc8f5c-47b1-453a-b1d9-aa9230630c86")
 
@@ -88,7 +83,6 @@ class EditMedicineEffectHandlerTest {
         patientUuid = patientUuid,
         facilityUuid = facilityUuid,
         syncStatus = SyncStatus.DONE,
-        timestamps = Timestamps.create(utcClock),
         frequency = MedicineFrequency.TDS,
         durationInDays = 30,
         teleconsultationId = UUID.fromString("d0e47d72-6773-4333-8447-bbd1c5004088")
@@ -103,55 +97,26 @@ class EditMedicineEffectHandlerTest {
         patientUuid = patientUuid,
         facilityUuid = facilityUuid,
         syncStatus = SyncStatus.DONE,
-        timestamps = Timestamps.create(utcClock),
         frequency = MedicineFrequency.TDS,
         durationInDays = 30,
         teleconsultationId = UUID.fromString("d0e47d72-6773-4333-8447-bbd1c5004088")
     )
     val prescriptions = listOf(drug1, drug2)
 
-    val clonedPrescriptions = listOf(
-        drug1.copy(
-            uuid = clonedDrug1Uuid,
-            facilityUuid = facility.uuid,
-            syncStatus = SyncStatus.PENDING,
-            timestamps = drug1.timestamps.copy(
-                createdAt = drug1.createdAt.plus(durationToAdvanceBy),
-                updatedAt = drug1.updatedAt.plus(durationToAdvanceBy),
-                deletedAt = null
-            ),
-            frequency = null,
-            durationInDays = null,
-            teleconsultationId = null
-        ),
-        drug2.copy(
-            uuid = clonedDrug2Uuid,
-            facilityUuid = facility.uuid,
-            syncStatus = SyncStatus.PENDING,
-            timestamps = drug2.timestamps.copy(
-                createdAt = drug2.createdAt.plus(durationToAdvanceBy),
-                updatedAt = drug2.updatedAt.plus(durationToAdvanceBy),
-                deletedAt = null
-            ),
-            frequency = null,
-            durationInDays = null,
-            teleconsultationId = null
-        )
-    )
-
     whenever(uuidGenerator.v4()).thenReturn(clonedDrug1Uuid, clonedDrug2Uuid)
     whenever(prescriptionRepository.newestPrescriptionsForPatientImmediate(patientUuid)) doReturn prescriptions
 
     // when
-    utcClock.advanceBy(durationToAdvanceBy)
     testCase.dispatch(RefillMedicines(patientUuid))
 
     // then
     testCase.assertOutgoingEvents(PrescribedMedicinesRefilled)
 
     verify(prescriptionRepository).newestPrescriptionsForPatientImmediate(patientUuid)
-    verify(prescriptionRepository).softDeletePrescriptions(prescriptions)
-    verify(prescriptionRepository).save(clonedPrescriptions)
+    verify(prescriptionRepository).refill(
+        prescriptions = eq(prescriptions),
+        uuidGenerator = any()
+    )
     verify(appointmentRepository).markAppointmentsCreatedBeforeTodayAsVisited(patientUuid)
     verifyNoMoreInteractions(prescriptionRepository)
     verifyNoMoreInteractions(appointmentRepository)