+++ app/src/main/java/org/simple/clinic/bp/history/BloodPressureHistoryViewEffectHandler.kt
@@ -11,6 +11,7 @@ class BloodPressureHistoryViewEffectHandler(
     when (viewEffect) {
       is OpenBloodPressureEntrySheet -> uiActions.openBloodPressureEntrySheet(viewEffect.patientUuid)
       is OpenBloodPressureUpdateSheet -> uiActions.openBloodPressureUpdateSheet(viewEffect.bloodPressureMeasurement.uuid)
+      is ShowBloodPressures -> uiActions.showBloodPressures(viewEffect.bloodPressureHistoryDataSourceFactory)
     }.exhaustive()
   }
 }