+++ app/src/test/java/org/simple/clinic/bp/history/BloodPressureHistoryScreenEffectHandlerTest.kt
@@ -35,7 +35,6 @@ class BloodPressureHistoryScreenEffectHandlerTest {
       patientRepository,
       TrampolineSchedulersProvider(),
       dataSourceFactory,
-      uiActions,
       viewEffectHandler::handle
   ).build()
   private val testCase = EffectHandlerTestCase(effectHandler)