+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientEffect.kt
@@ -10,38 +10,40 @@ data class LoadContactPatientProfile(val patientUuid: UUID) : ContactPatientEffe
 
 data class LoadLatestOverdueAppointment(val patientUuid: UUID) : ContactPatientEffect()
 
-data class DirectCallWithAutomaticDialer(val patientPhoneNumber: String) : ContactPatientEffect()
-
-data class DirectCallWithManualDialer(val patientPhoneNumber: String) : ContactPatientEffect()
-
 object LoadCurrentFacility : ContactPatientEffect()
 
-data class MaskedCallWithAutomaticDialer(
-    val patientPhoneNumber: String,
-    val proxyPhoneNumber: String
-) : ContactPatientEffect()
-
-data class MaskedCallWithManualDialer(
-    val patientPhoneNumber: String,
-    val proxyPhoneNumber: String
-) : ContactPatientEffect()
-
-object CloseScreen : ContactPatientEffect()
-
 data class MarkPatientAsAgreedToVisit(val appointment: Appointment) : ContactPatientEffect()
-
-data class ShowManualDatePicker(
-    val preselectedDate: LocalDate,
-    val datePickerBounds: ClosedRange<LocalDate>
-) : ContactPatientEffect()
 
 data class SetReminderForAppointment(
     val appointment: Appointment,
     val reminderDate: LocalDate
 ) : ContactPatientEffect()
 
+data class LoadCallResultForAppointment(val appointmentId: UUID) : ContactPatientEffect()
+
+sealed class ContactPatientViewEffect : ContactPatientEffect()
+
+data class DirectCallWithAutomaticDialer(val patientPhoneNumber: String) : ContactPatientViewEffect()
+
+data class DirectCallWithManualDialer(val patientPhoneNumber: String) : ContactPatientViewEffect()
+
+data class MaskedCallWithAutomaticDialer(
+    val patientPhoneNumber: String,
+    val proxyPhoneNumber: String
+) : ContactPatientViewEffect()
+
+data class MaskedCallWithManualDialer(
+    val patientPhoneNumber: String,
+    val proxyPhoneNumber: String
+) : ContactPatientViewEffect()
+
+object CloseScreen : ContactPatientViewEffect()
+
+data class ShowManualDatePicker(
+    val preselectedDate: LocalDate,
+    val datePickerBounds: ClosedRange<LocalDate>
+) : ContactPatientViewEffect()
+
 data class OpenRemoveOverdueAppointmentScreen(
     val appointment: Appointment
-) : ContactPatientEffect()
-
-data class LoadCallResultForAppointment(val appointmentId: UUID) : ContactPatientEffect()
+) : ContactPatientViewEffect()