+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientEffectHandler.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.contactpatient
 
+import com.spotify.mobius.functions.Consumer
 import com.spotify.mobius.rx2.RxMobius
 import dagger.Lazy
 import dagger.assisted.Assisted
@@ -11,7 +12,6 @@ import org.simple.clinic.facility.Facility
 import org.simple.clinic.overdue.AppointmentRepository
 import org.simple.clinic.overdue.callresult.CallResultRepository
 import org.simple.clinic.patient.PatientRepository
-import org.simple.clinic.phone.Dialer
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.scheduler.SchedulersProvider
 import java.time.LocalDate
@@ -25,12 +25,14 @@ class ContactPatientEffectHandler @AssistedInject constructor(
     private val schedulers: SchedulersProvider,
     private val currentFacility: Lazy<Facility>,
     private val callResultRepository: CallResultRepository,
-    @Assisted private val uiActions: ContactPatientUiActions
+    @Assisted private val viewEffectsConsumer: Consumer<ContactPatientViewEffect>
 ) {
 
   @AssistedFactory
   interface Factory {
-    fun create(uiActions: ContactPatientUiActions): ContactPatientEffectHandler
+    fun create(
+        viewEffectsConsumer: Consumer<ContactPatientViewEffect>
+    ): ContactPatientEffectHandler
   }
 
   fun build(): ObservableTransformer<ContactPatientEffect, ContactPatientEvent> {
@@ -38,17 +40,11 @@ class ContactPatientEffectHandler @AssistedInject constructor(
         .subtypeEffectHandler<ContactPatientEffect, ContactPatientEvent>()
         .addTransformer(LoadContactPatientProfile::class.java, loadContactPatientProfile(schedulers.io()))
         .addTransformer(LoadLatestOverdueAppointment::class.java, loadLatestOverdueAppointment(schedulers.io()))
-        .addConsumer(DirectCallWithAutomaticDialer::class.java, { uiActions.directlyCallPatient(it.patientPhoneNumber, Dialer.Automatic) }, schedulers.ui())
-        .addConsumer(DirectCallWithManualDialer::class.java, { uiActions.directlyCallPatient(it.patientPhoneNumber, Dialer.Manual) }, schedulers.ui())
-        .addConsumer(MaskedCallWithAutomaticDialer::class.java, { uiActions.maskedCallPatient(it.patientPhoneNumber, it.proxyPhoneNumber, Dialer.Automatic) }, schedulers.ui())
-        .addConsumer(MaskedCallWithManualDialer::class.java, { uiActions.maskedCallPatient(it.patientPhoneNumber, it.proxyPhoneNumber, Dialer.Manual) }, schedulers.ui())
-        .addAction(CloseScreen::class.java, uiActions::closeSheet, schedulers.ui())
         .addTransformer(MarkPatientAsAgreedToVisit::class.java, markPatientAsAgreedToVisit(schedulers.io()))
-        .addConsumer(ShowManualDatePicker::class.java, { uiActions.showManualDatePicker(it.preselectedDate, it.datePickerBounds) }, schedulers.ui())
         .addTransformer(SetReminderForAppointment::class.java, setReminderForAppointment(schedulers.io()))
-        .addConsumer(OpenRemoveOverdueAppointmentScreen::class.java, ::openRemoveOverdueAppointmentScreen, schedulers.ui())
         .addTransformer(LoadCurrentFacility::class.java, loadCurrentFacility())
         .addTransformer(LoadCallResultForAppointment::class.java, loadCallResultForAppointment())
+        .addConsumer(ContactPatientViewEffect::class.java, viewEffectsConsumer::accept)
         .build()
   }
 
@@ -68,10 +64,6 @@ class ContactPatientEffectHandler @AssistedInject constructor(
           .map { currentFacility.get() }
           .map(::CurrentFacilityLoaded)
     }
-  }
-
-  private fun openRemoveOverdueAppointmentScreen(effect: OpenRemoveOverdueAppointmentScreen) {
-    uiActions.openRemoveOverdueAppointmentScreen(effect.appointment)
   }
 
   private fun loadContactPatientProfile(