+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix select all not updating checkbox state in the overdue search screen
 - Bump Gradle to v7.5
 - Run prescription refill queries in a transaction
+- Use `PrescriptionRepository#refill` to refill prescriptions from edit medicines screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-08-01-8350