+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Run prescription refill queries in a transaction
 - Use `PrescriptionRepository#refill` to refill prescriptions from edit medicines screen
 - Use `PrescriptionRepository#refillForTeleconsulation` to refill prescriptions from teleconsult screen
+-  Migrate `BloodSugarHistoryScreen` to use view effects
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-08-01-8350